requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0
markdown==3.5.1
Pygments==2.16.1
